Fixing imports that are windows only
